Revamped table generator

+++ pipeline/split-segment.py
@@ -17,10 +17,10 @@ for t in types:
     if not os.path.exists(label_location):  
         os.mkdir(label_location)
 
-    data = images[:original_length/100 * t["range"]]
+    data = images[:int(original_length/100 * t["range"])]
 
     for image in data:
         os.rename(os.path.join(location, "train_A", image), os.path.join(img_location, image))
         os.rename(os.path.join(location, "train_B", image), os.path.join(label_location, image))
 
-    images = images[original_length/100 * t["range"])+1:]
+    images = images[int(original_length/100 * t["range"]+1):]
